#211 making a dosed pk model into a pkpd model
