+++ frontend/src/features/modelling/Chart.js
@@ -21,6 +21,8 @@ export default function ModellingChart({ datasets, pkModels, pkpdModels, pdModel
   const subjects = useSelector((state) => state.subjects.entities);
   const variables = useSelector((state) => state.variables.entities);
 
+  let labelsShown = {}
+
   const getChartData = (model) => {
     console.log("getChartData");
     if (!model.simulate) {
@@ -50,10 +52,18 @@ export default function ModellingChart({ datasets, pkModels, pkpdModels, pdModel
         if (variable.axis) {
           showRhsAxis = true;
         }
+        // chartjs crashes if there are identical labels
+        let label = variable.qname
+        if (label in labelsShown) {
+          label = `${label}${labelsShown[label]}`
+          labelsShown[label] += 1
+        } else {
+          labelsShown[label] = 1
+        }
         return {
           yAxisID: yAxisID,
           type: "line",
-          label: variable.name,
+          label: label,
           borderColor: color,
           backgroundColor: color,
           showLine: true,
@@ -116,7 +126,6 @@ export default function ModellingChart({ datasets, pkModels, pkpdModels, pdModel
       ...pkpdModels.map((m) => getChartData(m)).flat(),
     ],
   };
-  console.log('data', data)
 
   let options = {
     animation: {
@@ -175,6 +184,7 @@ export default function ModellingChart({ datasets, pkModels, pkpdModels, pdModel
     };
   }
 
+  console.log('data', data, options, renderChart)
   return (
     <div className={className}>
       {renderChart && <Scatter data={data} options={options} />}
